+++ src/packages/block/block-custom-view/index.ts
@@ -1,0 +1,1 @@
+export type * from './types.js';
