+++ src/packages/block/block-custom-view/types.ts
@@ -1,0 +1,36 @@
+import type { ManifestBlockEditorCustomView } from './block-editor-custom-view.extension.js';
+import type { UmbBlockLayoutBaseModel, UmbBlockDataType } from '@umbraco-cms/backoffice/block';
+import type { UmbBlockTypeBaseModel } from '@umbraco-cms/backoffice/block-type';
+
+export type * from './block-editor-custom-view.extension.js';
+
+export interface UmbBlockEditorCustomViewConfiguration {
+	editContentPath?: string;
+	editSettingsPath?: string;
+	showContentEdit: boolean;
+	showSettingsEdit: boolean;
+}
+
+export interface UmbBlockEditorCustomViewProperties<
+	LayoutType extends UmbBlockLayoutBaseModel = UmbBlockLayoutBaseModel,
+	BlockType extends UmbBlockTypeBaseModel = UmbBlockTypeBaseModel,
+> {
+	manifest?: ManifestBlockEditorCustomView;
+	config?: Partial<UmbBlockEditorCustomViewConfiguration>;
+	blockType?: BlockType;
+	contentUdi?: string;
+	label?: string;
+	icon?: string;
+	index?: number;
+	layout?: LayoutType;
+	content?: UmbBlockDataType;
+	settings?: UmbBlockDataType;
+	contentInvalid?: boolean;
+	settingsInvalid?: boolean;
+}
+
+export interface UmbBlockEditorCustomViewElement<
+	LayoutType extends UmbBlockLayoutBaseModel = UmbBlockLayoutBaseModel,
+	BlockType extends UmbBlockTypeBaseModel = UmbBlockTypeBaseModel,
+> extends UmbBlockEditorCustomViewProperties<LayoutType, BlockType>,
+		HTMLElement {}
