+++ src/packages/block/block-type/types.ts
@@ -1,4 +1,5 @@
 import type { UUIModalSidebarSize } from '@umbraco-cms/backoffice/external/uui';
+
 // Shared with the Property Editor
 export interface UmbBlockTypeBaseModel {
 	contentElementTypeKey: string;
